requirements.txt
asgiref==3.3.1
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-simplejwt==5.0.0
Pillow==8.0.1
PyJWT==2.3.0
pytz==2020.4
sqlparse==0.4.1
